haja coração

+++ GeneticMelody/GeneticMelody/Genetic/Mutation/MelodicReplacementOperator.cs
@@ -4,7 +4,7 @@ namespace GeneticMelody.Genetic.Mutation
 {
     public class MelodicReplacementOperator : IMelodyMutationOperator, IMutationOperator
     {
-        public double Rate => 10;
+        public int Rate => 10;
 
         public void Mutate(Melody melody)
         {
